feat: add api session check

+++ pages/api/comment/[id].ts
@@ -1,12 +1,15 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { getServerSession } from "next-auth/next"
+import { authOptions } from '../auth/[...nextauth]'
+
 import { Sequelize } from 'sequelize';
 import { Iarticle, Icomment, Iuser, errorResult, successResult } from '../../../type-config'
 import db from '../../../models/index';
 const DB: any = db;
 const { Users, Articles, Comments, Hollows, Likeships } = DB;
 
-export default function handleComments(req: NextApiRequest, res: NextApiResponse<successResult | errorResult>) {
+export default async function handleComments(req: NextApiRequest, res: NextApiResponse<successResult | errorResult>) {
     switch (req.method) {
         case 'PUT':
             editComment(req, res)
@@ -21,6 +24,9 @@ export default function handleComments(req: NextApiRequest, res: NextApiResponse
 }
 
 async function editComment (req: NextApiRequest, res: NextApiResponse<successResult | errorResult>) {
+    const session = await getServerSession(req, res, authOptions)
+    if (!session) return res.status(401).json({ error: '請先登入' })
+
     const { id } = req.query
     const idNum = Number(id)
     const { content } = req.body
@@ -32,6 +38,7 @@ async function editComment (req: NextApiRequest, res: NextApiResponse<successRes
     try {
         const comment = await Comments.findByPk(idNum, { transaction: t })
         if (!comment) return res.status(500).json({ error: '找不到該回覆' })
+        if (comment.user_id !== session.user.id) return res.status(401).json({ error: '使用者身分不符' })
 
         comment.set({ content }, { transaction: t })
         await comment.save({ transaction: t })
@@ -45,6 +52,9 @@ async function editComment (req: NextApiRequest, res: NextApiResponse<successRes
 }
 
 async function deleteComment (req: NextApiRequest, res: NextApiResponse<successResult | errorResult>) {
+    const session = await getServerSession(req, res, authOptions)
+    if (!session) return res.status(401).json({ error: '請先登入' })
+
     const { id } = req.query
     const idNum = Number(id)
     const t = await new Sequelize(process.env.MYSQL_DATABASE || '', process.env.MYSQL_USER || '', process.env.MYSQL_PASSWORD, {
@@ -52,12 +62,13 @@ async function deleteComment (req: NextApiRequest, res: NextApiResponse<successR
         dialect: 'mysql'
     }).transaction();
     try {
+        const comment = await Comments.findByPk(idNum, { transaction: t })
+        if (!comment) return res.status(500).json({ error: '此回覆不存在' })
+        if (comment.user_id !== session.user.id) return res.status(401).json({ error: '使用者身分不符' })
+        
         await Likeships.destroy({
             where: { comment_id: idNum }
         }, { transaction: t })
-
-        const comment = await Comments.findByPk(idNum, { transaction: t })
-        if (!comment) return res.status(500).json({ error: '此回覆不存在' })
 
         const article = await Articles.findByPk(comment.article_id, { transaction: t })
         if (article) {
